FLEXY-4614 Refactored files
