+++ .gitignore
@@ -145,8 +145,6 @@ test-results/
 
 appConfig.js
 
-package-lock.json
-
 # env files
 .env
 *.development
@@ -157,3 +155,5 @@ dependencies.csv
 
 #jUnit
 junit.xml
+
+.idea
