+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "plibo-call-recording-pause-and-resume",
-  "version": "0.0.0",
+  "version": "1.0.0",
   "private": true,
   "scripts": {
     "test": "jest",
@@ -12,15 +12,10 @@
     "@twilio-labs/serverless-runtime-types": "^2.2.3",
     "@twilio-labs/serverless-api": "^5.5.0",
     "lodash": "^4.17.21",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
     "twilio": "^4.0.0",
     "twilio-flex-token-validator": "^1.5.8"
   },
   "devDependencies": {
-    "@twilio-labs/runtime-helpers": "^0.1.2",
-    "@twilio-labs/serverless-api": "^5.5.0",
-    "@twilio-paste/core": "^17.0.1",
     "@babel/preset-env": "7.20.2",
     "@babel/preset-modules": "0.1.5",
     "@babel/preset-react": "7.18.6",
@@ -39,6 +34,6 @@
     "twilio-run": "3.4.2"
   },
   "engines": {
-    "node": "14"
+    "node": "16"
   }
 }
