+++ sonar-project.properties
@@ -1,5 +1,6 @@
 sonar.projectKey=twilio_flex-plugin-library-pause-recording
 sonar.organization=twilio
 
+sonar.exclusions=functions/setup.js,ui-src/**/**/redux/**/*,assets/**/*,ui-src/build/**/*,ui-src/**/utils/**/*,ui-src/**/**/strings/**/*,ui-src/**/**/states/**/*,ui-src/**/types/**/*,ui-src/test-utils/**/*,ui-src/public/**/*,coverage/**/*,**/*.index.tsx,**/*.index.ts,**/*.test.tsx,**/*.test.ts
 sonar.sources=ui-src/src,functions
 sonar.javascript.lcov.reportPaths=coverage/lcov.info
