+++ tests/helpers/prepare-function.private.test.js
@@ -1,5 +1,4 @@
-// import PrepareFunction from '../../../serverless/src/functions/helpers/prepare-function.private';
-import helpers from '../../test-utils/test-helper';
+import helpers from '../test-utils/test-helper';
 
 describe('Prepare Flex Function', () => {
   const mockFucntionValidatorObject = jest.fn();
