+++ ui-src/package.json
@@ -4,37 +4,19 @@
   "id": "plibo-call-recording-pause-and-resume",
   "private": true,
   "scripts": {
-    "test:watch": "jest --watch",
-    "test:mychanges": "jest --onlyChanged",
-    "test:precommit": "jest",
-    "test:ci": "jest --silent --ci --coverage --coverageReporters=cobertura --reporters=default --reporters=jest-junit --testResultsProcessor=jest-junit",
-    "test:coverage": "jest --silent --coverage && http-server -c-1 ./coverage -o /lcov-report",
-    "lint": "eslint --ext js --ext ts --ext tsx ./src/*.*",
-    "lint:fix": "npm run lint -- --fix",
-    "build": "twilio flex:plugins:build",
-    "deploy": "twilio flex:plugins:deploy --version=${npm_package_version}",
-    "install-flex-plugin": "twilio plugins:install @twilio-labs/plugin-flex@6.0.2",
-    "release": "twilio flex:plugins:release --plugin ${npm_package_name}@${npm_package_version}",
-    "start": "twilio flex:plugins:start",
-    "twilio": "twilio",
+    "prebuild": "rimraf build && npm run bootstrap",
+    "postinstall": "flex-plugin pre-script-check",
+    "prestart": "npm run bootstrap",
     "test": "jest"
   },
   "dependencies": {
-    "@testing-library/user-event": "14.4.3",
-    "@twilio-paste/core": "^17.0.1",
+    "@twilio-paste/core": "^17.2.0",
     "@twilio-paste/icons": "^9.4.2",
     "@twilio/flex-dev-utils": "^5.1.3",
-    "@twilio/flex-plugin-scripts": "6.0.3",
-    "@types/jest": "^26.0.20",
+    "@twilio/flex-plugin-scripts": "6.1.2",
     "lodash": "^4.17.20",
-    "luxon": "^3.1.1",
-    "prop-types": "^15.7.2",
     "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-select": "^5.5.0",
-    "rrule": "^2.7.1",
-    "twilio-video": "^2.24.0",
-    "tzdata": "^1.0.35"
+    "react-dom": "17.0.2"
   },
   "devDependencies": {
     "@testing-library/jest-dom": "5.16.5",
@@ -43,19 +25,11 @@
     "@babel/plugin-proposal-private-methods": "^7.12.1",
     "@babel/preset-typescript": "^7.12.1",
     "@twilio/flex-ui": "2.1.1",
-    "@types/luxon": "^3.1.0",
     "@types/react-redux": "^7.1.1",
-    "@types/uuid": "^8.3.4",
-    "@typescript-eslint/eslint-plugin": "^5.18.0",
-    "@typescript-eslint/parser": "^5.0.0",
-    "eslint": "^8.12.0",
-    "eslint-config-twilio": "^2.0.0",
     "jest": "^26.0.20",
-    "ts-jest": "26.5.6",
     "jest-fetch-mock": "3.0.3",
     "jest-junit": "12.0.0",
-    "react-test-renderer": "17.0.2",
-    "typescript": "^4",
+    "typescript": "^4.1.2",
     "twilio": "3.84.1",
     "twilio-flex-token-validator": "^1.5.4"
   },
